Move contents of IGameAction.h into GameAction.h
